requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cycler==0.11.0
defusedxml==0.7.1
Django==4.0
django-environ==0.8.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fonttools==4.28.5
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.0
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
xlwt==1.3.0
